Load the config file; #20
