+++ Cargo.toml
@@ -28,7 +28,9 @@ dialoguer = "0.11"
 time = "0.3"
 directories = "5.0"
 figment = { version = "0.10", features = ["toml"] }
+serde = { version = "1.0", features = ["derive"] }
 
 [build-dependencies]
 bpaf = { version = "0.9", features = ["derive", "docgen"]}
 time = "0.3"
+serde = { version = "1.0", features = ["derive"] }
