+++ src/cmd_line.rs
@@ -23,6 +23,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 use std::path::PathBuf;
 
 use bpaf::Bpaf;
+use serde::{Serialize, Deserialize};
 
 /// Generate pre-populated module files formatted with AsciiDoc that are used in Red Hat and Fedora documentation.
 #[derive(Clone, Debug, Bpaf)]
@@ -111,7 +112,7 @@ pub struct Action {
 
 /// The verbosity level set on the command line.
 /// The default option is invisible as a command-line argument.
-#[derive(Clone, Copy, Debug, Bpaf)]
+#[derive(Clone, Copy, Debug, Bpaf, Serialize, Deserialize)]
 pub enum Verbosity {
     /// Display additional, debug messages
     #[bpaf(short, long)]
