+++ src/lib.rs
@@ -33,8 +33,6 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 //! which are used in Red Hat and Fedora documentation. The generated files follow
 //! the Modular Documentation guidelines: <https://redhat-documentation.github.io/modular-docs/>.
 
-use std::path::PathBuf;
-
 use color_eyre::eyre::{bail, Result};
 
 pub mod cmd_line;
@@ -45,6 +43,7 @@ mod templating;
 mod write;
 
 use cmd_line::{Cli, Verbosity};
+pub use config::Options;
 pub use module::{ContentType, Input, Module};
 
 /// newdoc uses many regular expressions at several places. Constructing them should never fail,
@@ -52,52 +51,6 @@ pub use module::{ContentType, Input, Module};
 /// error message through `expect`.
 const REGEX_ERROR: &str = "Failed to construct a regular expression. Please report this as a bug";
 
-/// This struct stores options based on the command-line arguments,
-/// and is passed to various functions across the program.
-#[derive(Debug, Clone)]
-pub struct Options {
-    pub comments: bool,
-    pub file_prefixes: bool,
-    pub anchor_prefixes: bool,
-    pub examples: bool,
-    pub target_dir: PathBuf,
-    pub simplified: bool,
-    pub verbosity: Verbosity,
-}
-
-impl Options {
-    /// Set current options based on the command-line options
-    #[must_use]
-    pub fn new(cli: &Cli) -> Self {
-        Self {
-            // Comments and prefixes are enabled (true) by default unless you disable them
-            // on the command line. If the no-comments or no-prefixes option is passed,
-            // the feature is disabled, so the option is set to false.
-            comments: cli.common_options.comments,
-            file_prefixes: !cli.common_options.no_file_prefixes,
-            anchor_prefixes: cli.common_options.anchor_prefixes,
-            examples: !cli.common_options.no_examples,
-            // Set the target directory as specified or fall back on the current directory
-            target_dir: cli.common_options.target_dir.clone(),
-            simplified: cli.common_options.simplified,
-            verbosity: cli.common_options.verbosity,
-        }
-    }
-}
-
-impl Default for Options {
-    fn default() -> Self {
-        Self {
-            comments: false,
-            file_prefixes: true,
-            anchor_prefixes: false,
-            examples: true,
-            target_dir: PathBuf::from("."),
-            simplified: false,
-            verbosity: Verbosity::Default,
-        }
-    }
-}
 
 pub fn run(options: &Options, cli: &Cli) -> Result<()> {
     // Initialize the logging system based on the set verbosity
